This commit was generated by cvs2svn to compensate for changes in r5847, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@5848 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/utils.h
@@ -39,7 +39,7 @@ extern void *xmalloc(size_t s);
 extern void *xcalloc(size_t nmemb, size_t size);
 extern void *xrealloc(void *p, size_t s);
 extern char *concat(const char *s1, ...);
-extern char *bufcat(char *buf, int *buf_size, int *max_buf_size,
+extern char *bufcat(char *buf, int *buf_size, size_t *max_buf_size,
                     const char *src, int src_size);
 extern void remove_spaces(char *s);
 extern int make_backup_file(const char *fname);
